requirements.txt
flask
pyyaml
pandas
pytest
scikit-learn
mlflow
dvc